specified float for percetnages

+++ src/test/java/com/napier/sem/ReportUtilTest.java
@@ -22,8 +22,8 @@ public class ReportUtilTest
         Mockito.when(result.getLong("totalPopulation")).thenReturn(103000);
         Mockito.when(result.getLong("totalCity")).thenReturn(29034);
         Mockito.when(result.getInt("totalNotCity")).thenReturn(73966);
-        Mockito.when(result.getFloat("percentageCity")).thenReturn(28.19);
-        Mockito.when(result.getFloat("percentageNotCity")).thenReturn(71.81);
+        Mockito.when(result.getFloat("percentageCity")).thenReturn((float)28.19);
+        Mockito.when(result.getFloat("percentageNotCity")).thenReturn((float)71.81);
         ArrayList<Population> resultList = ReportUtil.peopleDistributionCountry(con);
         assertEquals(1, resultList.size());
         String expected = "Name: Aruba Population: 103000 City: 29034(28.19%) Not City: 73966(71.81%)";
